requirements.txt
ccxt==4.4.85
pandas==2.2.3
numpy==1.26.4
ta==0.11.0
python-dotenv==1.0.1
pydantic==2.11.3
pydantic-settings==2.8.1
loguru==0.7.3
python-telegram-bot==22.0
streamlit==1.44.1
PyYAML==6.0.2
pytest==8.3.5
